rest_api/requirements.txt
fastapi==0.90.0
mangum==0.17.0
pydantic==1.10.4
pynamodb==5.4.0
python-multipart==0.0.5
aws-lambda-powertools==2.7.0
websocket-client==1.5.0
python-dotenv==0.21.1